add exception handling to product scraper
