--- data/raw_data/v1_data/combo_data.json
@@ -1,24 +1,0 @@
-{
-  "combo_pack": [
-    {
-      "Monthly fee": "From 47,-",
-      "Data": "Unlimited",
-      "Download speed": "100Mbps",
-      "Upload speed": "40Mbps",
-      "Mobile subscription": "1 GB to 70 GB",
-      "Technology": "Copper",
-      "combo_advantage": "3",
-      "name": "Unlimited fast internet"
-    },
-    {
-      "Monthly fee": "From 62,-",
-      "Data": "Unlimited",
-      "Download speed": "1Gbps",
-      "Upload speed": "500Mbps",
-      "Mobile subscription": "1 GB to 70 GB",
-      "Technology": "Fiber",
-      "combo_advantage": "3",
-      "name": "Unlimited superfast internet"
-    }
-  ]
-}
