--- data/raw_data/v1_data/internet_data.json
@@ -1,20 +1,0 @@
-{
-    "internet_plans": [
-        {
-            "Monthly fee": "40,-",
-            "Data": "Unlimited",
-            "Download speed": "100Mbps",
-            "Upload speed": "40Mbps",
-            "Technology": "Copper",
-            "name": "Unlimited fast internet"
-        },
-        {
-            "Monthly fee": "55,-",
-            "Data": "Unlimited",
-            "Download speed": "1Gbps",
-            "Upload speed": "500Mbps",
-            "Technology": "Fiber",
-            "name": "Unlimited superfast internet"
-        }
-    ]
-}
