--- data/raw_data/v1_data/subscription_data.json
@@ -1,44 +1,0 @@
-{
-    "subscription_plans": [
-        {
-            "id": 1,
-            "price_per_month": "10",
-            "mobile_data": "2 gb",
-            "network": "4g",
-            "minutes": "150",
-            "sms": "unlimited"
-        },
-        {
-            "id": 2,
-            "price_per_month": "12",
-            "mobile_data": "4 gb",
-            "network": "4g",
-            "minutes": "150",
-            "sms": "unlimited"
-        },
-        {
-            "id": 3,
-            "price_per_month": "15",
-            "mobile_data": "10 gb",
-            "network": "5g",
-            "minutes": "unlimited",
-            "sms": "unlimited"
-        },
-        {
-            "id": 4,
-            "price_per_month": "20",
-            "mobile_data": "20 gb",
-            "network": "5g",
-            "minutes": "unlimited",
-            "sms": "unlimited"
-        },
-        {
-            "id": 5,
-            "price_per_month": "29",
-            "mobile_data": "70 gb",
-            "network": "5g",
-            "minutes": "unlimited",
-            "sms": "unlimited"
-        }
-    ]
-}
